+++ src/Core/Tags/WebScrap.Core.Tags/Tools/Extracting/AttributeExtractor.cs
@@ -1,4 +1,4 @@
-namespace WebScrap.Core.Tags.Extracting;
+namespace WebScrap.Core.Tags.Tools.Extracting;
 
 internal static class AttributeExtractor
 {
